Use `Test::MockTime` during testing.
There was a recent CPANTesters failure, due to a race condition in the
test between a call to `time()` and the generation of template output.

Both were making their own calls to `time()`, and between the two calls
we slipped into the next second, thus causing a discrepancy between the
values.

So... set a fixed time for the duration of the test using
`Test::MockTime`, so that "time effectively stands still" during the
test.